serving requests, network error fixed

+++ patashambaBackend/chat/models.py
@@ -5,7 +5,7 @@ from users.models import user
 # Create your models here.
 
 class chat_room(models.Model):
-    room_id = models.BigIntegerField(primary_key = True)
+    room_id = models.AutoField(primary_key=True)
     land_id = models.ForeignKey(land, on_delete=models.CASCADE, null=True)
     room_open = models.BooleanField(default=False)
